glep-0068: Add <stabilize-allarches/> element to metadata.xml
Add a new <stabilize-allarches/> element to metadata.xml whose presence
can be used to indicate that the package is suitable for ALLARCHES
stabilization. Unlike the current system, this makes it possible to
easily determine when packages can be stabilized wrt ALLARCHES
independently of who requests the stabilization.

For example, in Python ecosystem there are some packages which have C
extensions or known-unportable code which we do not want to handle
through ALLARCHES. Right now we have to either remember them or recheck
for them when firing the stablereq, or — more commonly — we end up
not using ALLARCHES when we could.

Storing this information per-package will not only make it easier to
determine whether ALLARCHES is applicable but also introduce some
automation to determine that and react accordingly.

Approved on 2018-03-11 Council meeting.

Closes: https://bugs.gentoo.org/649740