glep-0063: Require encryption subkey, and make primary certify-only
Following the recent mailing list discussion indicating that developers
are taking GLEP 63 as only source of truth about OpenPGP keys, and can
make assumption that if encryption key is not listed there they should
not have one.  Amend the specification to extend it beyond the previous
limited scope of commit signing, and require an encryption key
appropriately.  This matches the GnuPG defaults.

While at it, add a recommendation that the primary key is certify-only.
Other usage is implicitly discouraged anyway via requiring subkeys.
Originally this recommendation was omitted as I wasn't aware that gpg
had a (hidden) option to change usage of existing keys.

Closes: https://bugs.gentoo.org/681802
Signed-off-by: Michał Górny <mgorny@gentoo.org>