glep-0082: New key "eapis-testing"
As discussed in #gentoo-portage yesterday: After a new EAPI has been
approved, there is a period during which there is no stable Portage
version supporting the new EAPI, and ebuilds for the new EAPI should
not be stabilised during that period. Add a new "eapis-testing" key
to layout.conf, so that development tools can warn about it.

Suggested-by: Michał Górny <mgorny@gentoo.org>
Signed-off-by: Ulrich Müller <ulm@gentoo.org>