glep-0076: Relax name policy to allow pseudonyms
Update the signoff rules of GLEP 76 to allow using established
pseudonyms in addition to legal names.  This follows a similar change
that has been carried in the Linux kernel's documentation [1] that
Gentoo's policy has been based on.

This change is proposed instead of the "version 1.2" that has been
approved on 2022-12-11 but have not been pushed yet.  That version made
the wording significantly more complex without actually addressing
the original issue.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=d4563201f33a022fc0353033d9dfeb1606a88330

Bug: https://bugs.gentoo.org/900857
Signed-off-by: Michał Górny <mgorny@gentoo.org>
Signed-off-by: Ulrich Müller <ulm@gentoo.org>